moved the compression to upload, so quality is kept for local viz and download
